# apispec.yaml
openapi: 3.0.2
info:
  title: Reviews API
  version: "1.0"
  contact:
    name: My name
    email: email@example.com
  description: This API allows user to manage reviews
servers:
  - url: "http://localhost:5001/api"
    description: Development server
paths:
  /reviews:
    get:
      description: Get a list of reviews
      operationId: get-reviews
      tags:
        - reviews
      parameters:
        - schema:
            type: string
          in: query
          name: style
          description: The style of ramen noodles
          required: false
        - schema:
            type: string
          in: query
          name: brands
          description: Comma separated list of brands of ramen noodles
          required: false
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
        500:
          description: Internal Server Error
    post:
      description: Create a new review
      operationId: post-reviews
      tags:
        - reviews
      responses:
        201:
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
        400:
          description: Bad Request
        500:
          description: Internal Server Error
  /reviews/{reviewId}:
    parameters:
      - schema:
          type: string
        name: reviewId
        in: path
        required: true
    get:
      description: Retrieve the information of the review with the matching review ID.
      operationId: get-reviews-reviewId
      tags:
        - reviews
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
        404:
          description: Not Found
        500:
          description: Internal Server Error
    patch:
      description: Update an existing review.
      operationId: patch-reviews-reviewId
      tags:
        - reviews
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
        404:
          description: Not Found
        500:
          description: Internal Server Error
    delete:
      description: Deletes a specific review
      operationId: delete-reviews-reviewId
      tags:
        - reviews
      responses:
        200:
          description: OK
        404:
          description: Not Found
        500:
          description: Internal Server Error
components:
  schemas:
    Review:
      type: object
      properties:
        id:
          type: integer
        url:
          type: string
        brand:
          type: string
        variety:
          type: string
        packagingStyle:
          type: string
        country:
          type: string
        rating:
          type: number
      required:
        - brand
        - variety
        - packagingStyle
        - country
        - rating
tags:
  - name: reviews